mv api::NetAddress, setting into kld
