+++ Cargo.toml
@@ -2,6 +2,5 @@
 members = [
     "api",
     "kld",
-    "settings",
     "test-utils",
 ]
