+++ api/Cargo.toml
@@ -8,11 +8,7 @@ anyhow = { version = "1.0.71", features = [ "backtrace" ] }
 bitcoin = { version = "0.29.2", features = [ "serde" ] }
 serde = { version = "1.0", features = [ "derive" ] }
 serde_json = "1.0"
-lightning = { version = "0.0.115", features = ["max_level_trace", "_test_utils"] }
 hex = "0.4"
-
-[dev-dependencies]
-bincode = "1.3.3"
 
 [lib]
 doctest = false
