+++ kld/src/api/payments.rs
@@ -1,9 +1,9 @@
 use std::{str::FromStr, sync::Arc, time::UNIX_EPOCH};
 
-use api::lightning::routing::gossip::NodeId;
 use api::{KeysendRequest, ListPaysParams, PayInvoice, Payment, PaymentResponse};
 use axum::{extract::Query, response::IntoResponse, Extension, Json};
 use hex::ToHex;
+use lightning::routing::gossip::NodeId;
 
 use crate::{database::invoice::Invoice, ldk::LightningInterface};
