+++ kld/src/bitcoind/bitcoind_client.rs
@@ -8,13 +8,12 @@ use std::{
 
 use anyhow::{anyhow, bail, Context, Result};
 
-use api::lightning::chain::chaininterface::{
-    BroadcasterInterface, ConfirmationTarget, FeeEstimator,
-};
+use crate::settings::Settings;
 use async_trait::async_trait;
 use base64::{engine::general_purpose, Engine};
 use bitcoin::{consensus::encode, Address, BlockHash, Transaction, Txid};
 use bitcoincore_rpc_json::{EstimateMode, EstimateSmartFeeResult, GetBlockchainInfoResult};
+use lightning::chain::chaininterface::{BroadcasterInterface, ConfirmationTarget, FeeEstimator};
 use lightning_block_sync::{
     http::{HttpEndpoint, JsonResponse},
     rpc::RpcClient,
@@ -23,7 +22,6 @@ use lightning_block_sync::{
 use log::{error, info};
 use serde::Deserialize;
 use serde_json::{json, Value};
-use settings::Settings;
 use tokio::runtime::Handle;
 
 use crate::{ldk::MIN_FEERATE, quit_signal, Service};
