+++ kld/src/bitcoind/utxo_lookup.rs
@@ -1,14 +1,14 @@
 use std::sync::{Arc, Weak};
 
 use crate::logger::KldLogger;
-use api::lightning::routing::{
+use crate::settings::Settings;
+use bitcoin::{blockdata::constants::genesis_block, BlockHash};
+use lightning::routing::{
     gossip::P2PGossipSync,
     utxo::{UtxoFuture, UtxoLookup, UtxoLookupError, UtxoResult},
 };
-use bitcoin::{blockdata::constants::genesis_block, BlockHash};
 use lightning_block_sync::{BlockData, BlockSource};
 use log::warn;
-use settings::Settings;
 use tokio::runtime::Handle;
 
 use crate::ldk::{
