+++ kld/src/wallet/bdk_wallet.rs
@@ -4,10 +4,8 @@ use std::{
     time::{Duration, SystemTime, UNIX_EPOCH},
 };
 
+use crate::settings::{Network, Settings};
 use anyhow::{bail, Result};
-use api::lightning::chain::chaininterface::{
-    BroadcasterInterface, ConfirmationTarget, FeeEstimator,
-};
 use async_trait::async_trait;
 use bdk::{
     bitcoin::util::bip32::ExtendedPrivKey,
@@ -25,10 +23,10 @@ use bitcoin::{
     util::bip32::{ChildNumber, DerivationPath},
     Address, OutPoint, Script, Transaction,
 };
+use lightning::chain::chaininterface::{BroadcasterInterface, ConfirmationTarget, FeeEstimator};
 use lightning_block_sync::BlockSource;
 use log::{error, info, warn};
 use once_cell::sync::OnceCell;
-use settings::{Network, Settings};
 
 use crate::Service;
 
@@ -333,11 +331,11 @@ mod test {
         sync::{Arc, Mutex},
     };
 
+    use crate::settings::Settings;
     use anyhow::Result;
     use bdk::{database::MemoryDatabase, wallet::get_funded_wallet, Balance};
     use bitcoin::Address;
     use once_cell::sync::OnceCell;
-    use settings::Settings;
     use test_utils::{TEST_ADDRESS, TEST_WPKH};
 
     use crate::{bitcoind::MockBitcoindClient, wallet::WalletInterface};
