+++ kld/tests/database/ldk_database.rs
@@ -10,18 +10,18 @@ use kld::database::payment::{Payment, PaymentDirection, PaymentStatus};
 use kld::database::peer::Peer;
 use kld::database::LdkDatabase;
 
-use api::lightning::chain::chaininterface::{BroadcasterInterface, FeeEstimator};
-use api::lightning::chain::chainmonitor::ChainMonitor;
-use api::lightning::chain::keysinterface::{InMemorySigner, KeysManager};
-use api::lightning::chain::Filter;
-use api::lightning::ln::channelmanager::PaymentId;
-use api::lightning::ln::msgs::NetAddress;
-use api::lightning::ln::{PaymentPreimage, PaymentSecret};
-use api::lightning::routing::gossip::{NetworkGraph, NodeId};
-use api::lightning::routing::router::DefaultRouter;
-use api::lightning::routing::scoring::{ProbabilisticScorer, ProbabilisticScoringParameters};
-use api::lightning::util::persist::Persister;
 use kld::logger::KldLogger;
+use lightning::chain::chaininterface::{BroadcasterInterface, FeeEstimator};
+use lightning::chain::chainmonitor::ChainMonitor;
+use lightning::chain::keysinterface::{InMemorySigner, KeysManager};
+use lightning::chain::Filter;
+use lightning::ln::channelmanager::PaymentId;
+use lightning::ln::msgs::NetAddress;
+use lightning::ln::{PaymentPreimage, PaymentSecret};
+use lightning::routing::gossip::{NetworkGraph, NodeId};
+use lightning::routing::router::DefaultRouter;
+use lightning::routing::scoring::{ProbabilisticScorer, ProbabilisticScoringParameters};
+use lightning::util::persist::Persister;
 use lightning_invoice::{Currency, InvoiceBuilder};
 use rand::random;
 use test_utils::{poll, random_public_key, TEST_PRIVATE_KEY};
