+++ kld/tests/database/main.rs
@@ -9,10 +9,10 @@ use futures::Future;
 use futures::FutureExt;
 use kld::database::DurableConnection;
 use kld::logger::KldLogger;
+use kld::settings::Settings;
 use kld::Service;
 use once_cell::sync::Lazy;
 use once_cell::sync::OnceCell;
-use settings::Settings;
 use test_utils::cockroach_manager::create_database;
 use test_utils::poll;
 use test_utils::test_settings;
