--- settings/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "settings"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-once_cell = "1.18.0"
-clap = { version = "4.3.1", features = ["derive", "env"] }
-bitcoin = "0.29.2"
-api = { path = "../api" }
-
-[lib]
-doctest = false
